requirements.txt
fastapi
uvicorn
groq
pydantic
httpx
python-multipart
python-dotenv
